+++ src/main/java/com/skycat/mystical/Mystical.java
@@ -18,6 +18,7 @@ import net.fabricmc.fabric.api.event.lifecycle.v1.ServerWorldEvents;
 import net.fabricmc.fabric.api.event.player.AttackBlockCallback;
 import net.fabricmc.fabric.api.event.player.PlayerBlockBreakEvents;
 import net.minecraft.advancement.criterion.Criteria;
+import net.minecraft.advancement.criterion.ItemCriterion;
 import net.minecraft.block.Block;
 import net.minecraft.entity.EntityType;
 import net.minecraft.registry.RegistryKeys;
@@ -31,7 +32,7 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Random;
 
-// WARNING: Package structure looks like split-sources structure. It's not. I'm working on it.
+// WARNING: Package structure in common looks like split-sources structure, which is weird. I'm working on it.
 public class Mystical implements ModInitializer, ServerWorldEvents.Load {
     public static final String MOD_ID = "mystical"; // TODO: Use this everywhere it makes sense
     public static final Logger LOGGER = LoggerFactory.getLogger("mystical");
@@ -49,7 +50,8 @@ public class Mystical implements ModInitializer, ServerWorldEvents.Load {
     public static final TagKey<Block> GLAZED_TERRACOTTA = TagKey.of(RegistryKeys.BLOCK, new Identifier("mystical:glazed_terracotta"));
     public static final MakeHavenCriterion MAKE_HAVEN_CRITERION = Criteria.register(new MakeHavenCriterion());
     public static final SpellCuredCriterion SPELL_CURED_CRITERION = Criteria.register(new SpellCuredCriterion());
-
+    public static final Identifier PREVENTED_BREAKING_CRITERION_ID = Identifier.of(Mystical.MOD_ID, "prevented_breaking");
+    public static final ItemCriterion PREVENTED_BREAKING_CRITERION = Criteria.register(new ItemCriterion(PREVENTED_BREAKING_CRITERION_ID));
     public static SaveState save;
     private static boolean isClientWorld = true;